+++ src/OmniSharp.Abstractions/Models/v2/BlockStructureRequest.cs
@@ -1,6 +1,6 @@
 ﻿using OmniSharp.Mef;
 
-namespace OmniSharp.Models.v2
+namespace OmniSharp.Models.V2
 {
     [OmniSharpEndpoint(OmniSharpEndpoints.V2.BlockStructure, typeof(BlockStructureRequest), typeof(BlockStructureResponse))]
     public class BlockStructureRequest : SimpleFileRequest