+++ src/OmniSharp.MSBuild/ProjectSystem.cs
@@ -10,6 +10,7 @@ using Microsoft.Extensions.Logging;
 using OmniSharp.Eventing;
 using OmniSharp.FileSystem;
 using OmniSharp.FileWatching;
+using OmniSharp.Mef;
 using OmniSharp.Models.WorkspaceInformation;
 using OmniSharp.MSBuild.Discovery;
 using OmniSharp.MSBuild.Models;
@@ -20,7 +21,7 @@ using OmniSharp.Services;
 
 namespace OmniSharp.MSBuild
 {
-    [Export(typeof(IProjectSystem)), Shared]
+    [ExportProjectSystem(ProjectSystemNames.MSBuildProjectSystem), Shared]
     public class ProjectSystem : IProjectSystem
     {
         private readonly IOmniSharpEnvironment _environment;