+++ src/OmniSharp.Roslyn/OmniSharpWorkspace.cs
@@ -1,4 +1,5 @@
 using System;
+using System.Collections.Concurrent;
 using System.Collections.Generic;
 using System.Composition;
 using System.IO;
@@ -8,7 +9,9 @@ using System.Threading;
 using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.Text;
 using Microsoft.Extensions.Logging;
+using OmniSharp.FileWatching;
 using OmniSharp.Roslyn;
+using OmniSharp.Roslyn.Utilities;
 using OmniSharp.Utilities;
 
 namespace OmniSharp
@@ -21,11 +24,13 @@ namespace OmniSharp
 
         private readonly ILogger<OmniSharpWorkspace> _logger;
 
+        private readonly ConcurrentDictionary<string, ProjectInfo> miscDocumentsProjectInfos = new ConcurrentDictionary<string, ProjectInfo>();
+
         [ImportingConstructor]
-        public OmniSharpWorkspace(HostServicesAggregator aggregator, ILoggerFactory loggerFactory)
+        public OmniSharpWorkspace(HostServicesAggregator aggregator, ILoggerFactory loggerFactory, IFileSystemWatcher fileSystemWatcher)
             : base(aggregator.CreateHostServices(), "Custom")
         {
-            BufferManager = new BufferManager(this);
+            BufferManager = new BufferManager(this, fileSystemWatcher);
             _logger = loggerFactory.CreateLogger<OmniSharpWorkspace>();
         }
 
@@ -80,7 +85,48 @@ namespace OmniSharp
 
         public void AddDocument(DocumentInfo documentInfo)
         {
+            // if the file has already been added as a misc file,
+            // because of a possible race condition between the updates of the project systems,
+            // remove the misc file and add the document as required
+            TryRemoveMiscellaneousDocument(documentInfo.FilePath);
+
             OnDocumentAdded(documentInfo);
+        }
+
+        public DocumentId TryAddMiscellaneousDocument(string filePath, string language)
+        {
+            if (GetDocument(filePath) != null)
+                return null; //if the workspace already knows about this document then it is not a miscellaneous document
+
+            var projectInfo = miscDocumentsProjectInfos.GetOrAdd(language, (lang) => CreateMiscFilesProject(lang));
+            var documentId = AddDocument(projectInfo.Id, filePath);
+            return documentId;
+        }
+
+        public bool TryRemoveMiscellaneousDocument(string filePath)
+        {
+            var documentId = GetDocumentId(filePath);
+            if (documentId == null || !IsMiscellaneousDocument(documentId))
+                return false;
+
+            RemoveDocument(documentId);
+            return true;
+        }
+
+        private ProjectInfo CreateMiscFilesProject(string language)
+        {
+            string assemblyName = Guid.NewGuid().ToString("N");
+            var projectInfo = ProjectInfo.Create(
+                   id: ProjectId.CreateNewId(),
+                   version: VersionStamp.Create(),
+                   name: "MiscellaneousFiles",
+                   metadataReferences: DefaultMetadataReferenceHelper.GetDefaultMetadataReferenceLocations()
+                                       .Select(loc => MetadataReference.CreateFromFile(loc)),
+                   assemblyName: assemblyName,
+                   language: language);
+
+            AddProject(projectInfo);
+            return projectInfo;
         }
 
         public DocumentId AddDocument(ProjectId projectId, string filePath, SourceCodeKind sourceCodeKind = SourceCodeKind.Regular)
@@ -223,6 +269,16 @@ namespace OmniSharp
             {
                 _logger.LogError(e, $"Error saving document {fullPath}");
             }
+        }
+
+        public bool IsCapableOfSemanticDiagnostics(Document document)
+        {
+            return !IsMiscellaneousDocument(document.Id);
+        }
+
+        private bool IsMiscellaneousDocument(DocumentId documentId)
+        {
+            return miscDocumentsProjectInfos.Where(p => p.Value.Id == documentId.ProjectId).Any();
         }
     }
 }